requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
dnspython==2.5.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.2
greenlet==3.0.3
h11==0.14.0
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
python-decouple==3.8
python-stdnum==1.19
sniffio==1.3.0
SQLAlchemy==2.0.27
sqlmodel==0.0.14
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.1
